Issue #611 eliminate unnecessary `Capabilties` base class
